Break TurnToward but fix rotation and y-inversion
Now origin 0,0 is at bottom left corner and inversion of y coordinate is
only done during drawing stage to project onto canvas. Effectively the
simulation must be mirrored on the y axis to be drawn because
historically graphics are drawn from top to bottom. That's kind of
stupid when you think about it.